Sim Controller Tuning
Corresponds with RethinkRobotics/sawyer_simulator#35

Now that all the controllers have been created, and the software
processes are in place to emulate the real Sawyer robot's API's, this is
the latest round of tuning for Sawyer's Position(&Trajectory) and
Velocity Controllers. This corresponds with some URDF tweaks in order to
make Gazebo happy: increased zero mass & inertia links to have a minimum
mass and inertia, and re-tuned joint damping and friction for
simulation.

This was tested with a computer capable of running Sawyer in Gazebo at a
consistent Real Time Factor of 1.0 throughout the duration of testing.

    Trajectory Mode works well(-ish) with MoveIt! Nearly all
trajectories execute without violating the real Sawyer's joint tracking
or Goal Velocity constraints
    Regular Position Mode moves at nearly the same speed as the real
Sawyer
    Velocity Mode works well with the joint_velocity_wobbler.py example
script

The one slight downside is that with the retuning, and mass/inertia
tweaks, Gravity Compensation is just barely not enough to hold up the
arm when commanding only Zero Torques. Funny enough, this is actually
closer to the reality of the real Sawyer.